UV_Streamlit/pyproject.toml
[project]
name = "uv-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.18",
    "langchain-google-genai>=2.0.9",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.0",
]

[project.scripts]
uv-streamlit = "uv_streamlit:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
